requirements.txt
llama-cpp-python
langchain>=0.2.2
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
PyMuPDF
python-docx
openpyxl
PyYAML
pandas
tqdm
streamlit
matplotlib
seaborn